redir
open de los archivos de las redirecciones hecho.

+++ src/exeggcute/exeggutor.c
@@ -6,7 +6,7 @@
 /*   By: eralonso <eralonso@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/13 13:56:31 by eralonso          #+#    #+#             */
-/*   Updated: 2023/06/17 17:40:37 by eralonso         ###   ########.fr       */
+/*   Updated: 2023/06/18 12:58:47 by eralonso         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,13 @@ int	kill_childs(pid_t *pids, int size)
 
 	i = -1;
 	while (++i < size)
-		kill(pids[i], SIGTERM);
+	{
+		if (pids[i] > 0)
+		{
+			kill(pids[i], SIGTERM);
+			pids[i] = 0;
+		}
+	}
 	return (1);
 }
 
@@ -37,7 +43,7 @@ int	wait_childs(pid_t *pids, int size)
 	{
 		aux = waitpid(-1, &status, 0);
 		if (aux == -1)
-			return (kill_childs(pids, size));
+			return (1);
 		if (WIFEXITED(status) && aux == pids[size - 1])
 			(1 && (last = 1) && (g_msh.err = WEXITSTATUS(status)));
 		else if (WIFSIGNALED(status))
@@ -57,6 +63,11 @@ pid_t	exec_node(t_lstt *node, int idx, int end)
 
 	if (redirect_parser(node->redirect, node->redir_size))
 		return (ERR_NODE);
+	if (redirect_node(node))
+		return (ERR_NODE);	
+	child = 0;
+	(void) idx;
+	(void) end;
 	// redirect_parser(&node);
 	// do_redirections(&node);
 	// init_signals(N_INTERACT);
@@ -73,50 +84,51 @@ pid_t	exec_node(t_lstt *node, int idx, int end)
 	// 	return (child);
 	// }
 	// exapand_commands();
-	if (idx == 0 && end && is_builtin(((t_cmd *)node->content)->cmd_args[0]))
-		return (1);
+	// if (idx == 0 && end && is_builtin(((t_cmd *)node->content)->cmd_args[0]))
+	// 	return (1);
 		// return (exec_builtins(node->content));
 	//Si exec node no tiene que hacer fork, devuelve 0.
 	// Devuelve ERR_NODE en caso de fallar
-	child = fork();
-	if (child < 0)
-		return (ERR_NODE);
+	// child = fork();
+	// if (child < 0)
+	// 	return (ERR_NODE);
 	// if (child == 0)
 	// 	exec_cmd(node->content);
 	return (child);
 }
 
-int	exec_nodes(t_lstt **node, int size)
+int	exec_nodes(t_lstt **node, int size, const int std_fd[2])
 {
-	t_lstt	*tmp;
-	pid_t	*pids;
-	int		i;
-	int		tmp_fd[2];
+	t_lstt		*tmp;
+	pid_t		*pids;
+	int			i;
+	int			tmp_fd[2];
 
-	if (!node || !*node)
-		return (1);
 	pids = ft_calloc(sizeof(pid_t), size);
-	if (!pids)
+	if (!node || !*node || !pids)
 		return (1);
 	(1 && (tmp = *node) && (i = 0));
-	(1 && (tmp_fd[0] = dup(0)) && (tmp_fd[1] = dup(1)));
+	(1 && (tmp_fd[0] = dup(std_fd[0])) && (tmp_fd[1] = dup(std_fd[1])));
+	if (tmp_fd[0] == -1 || tmp_fd[1] == -1)
+		return (free(pids), 1);
 	while (tmp)
 	{
+		if (pipe(tmp->fd) == -1)
+			return (kill_childs(pids, size), free(pids), 1);
 		pids[i] = exec_node(tmp, i, i == (size - 1));
 		if (pids[i++] == ERR_NODE)
 			return (kill_childs(pids, size), free(pids), 1);
 		tmp = tmp->next;
 	}
-	dup2(tmp_fd[0], 0);
-	ft_close(&tmp_fd[0]);
-	dup2(tmp_fd[1], 1);
-	ft_close(&tmp_fd[1]);
-	i = wait_childs(pids, size);
+	(((redir_std(tmp_fd, std_fd) || wait_childs(pids, size)) \
+	&& (i = kill_childs(pids, size))) || (i = 0));
 	return (free(pids), i);
 }
 
 int	executor(t_stair *st)
 {
+	const int	std_fd[2] = {0, 1};
+
 	if (!st)
 		return (0);
 	if (st->type != MAIN)
@@ -124,7 +136,7 @@ int	executor(t_stair *st)
 			return (1);
 	if ((st->type == AND && g_msh.err) || (st->type == OR && !g_msh.err))
 		return (0);
-	if (exec_nodes(&st->node, st->size))
+	if (exec_nodes(&st->node, st->size, std_fd))
 		return (1);
 	return (0);
 }
